requirements.txt
amqp==5.3.1
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.5
django-cors-headers==4.6.0
django-debug-toolbar==5.0.1
django-filter==24.3
django-simple-history==3.8.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
Markdown==3.7
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
